+++ R/raster-input.R
@@ -110,12 +110,18 @@ vapour_read_raster <- function(x, band = 1, window, resample = "nearestneighbour
 #' separate from propagated missing "NODATA" values in the source.
 #'
 #' Argument 'source_wkt' may be used to assign the projection of the source, 'source_extent'
-#' to assign the extent of the source. Sometimes both are required. Wild combinations of
+#' to assign the extent of the source. Sometimes both are required.
+#'
+#' If multiple sources are specified via 'x' and either 'source_wkt' or 'source_extent' are provided, these
+#' are applied to every source even if they have valid values already. If this is not sensible please use VRT
+#' to wrap the multiple sources first (see the gdalio package for some in-dev ideas).
+#'
+#' Wild combinations of
 #' 'source_extent' and/or 'extent' may be used for arbitrary flip orientations, scale and offset. For
 #' expert usage only. Old versions allowed transform input for target and source but this is now disabled (maybe we'll write
 #'  a new wrapper for that).
 #'
-#' @param x data source string (file name or URL or database connection string)
+#' @param x vector of data source names (file name or URL or database connection string)
 #' @param bands index of band/s to read (1-based), may be new order or replicated, or NULL (all bands used)
 #' @param extent extent of the target warped raster 'c(xmin, xmax, ymin, ymax)'
 #' @param source_extent extent of the source raster, used to override/augment incorrect source metadata
@@ -244,7 +250,7 @@ vapour_warp_raster <- function(x, bands = 1L,
     resample <- "near"
   }
   rso <- c("near", "bilinear", "cubic", "cubicspline", "lanczos", "average",
-           "mode", "max", "min", "med", "q1", "q3") ## "sum", "rms")
+           "mode", "max", "min", "med", "q1", "q3", "sum") #, "rms")
 
   if (!resample %in% rso) {
     warning(sprintf("%s resampling not available for warper, using near", resample))
@@ -255,10 +261,10 @@ vapour_warp_raster <- function(x, bands = 1L,
 
   vals <- warp_in_memory_gdal_cpp(x, source_WKT = source_wkt,
                                    target_WKT = wkt,
-                                   target_extent = extent,
-                                   target_dim = dimension,
-                                  bands = bands,
-                                  source_extent = source_extent,
+                                   target_extent = as.numeric(extent),
+                                   target_dim = as.integer(dimension),
+                                  bands = as.integer(bands),
+                                  source_extent = as.numeric(source_extent),
                                   resample = resample,
                                   silent = silent)
   if (length(bands) == 1 && bands == 0) {